Moved test files to "tests" folder, renamed send_unicorn to unicorn
